Create Dev Container suitable for using the automl-embedded extension
This commit creates a new Dev Container definition and and associated
Dockerfile and west.yml file. The Dev Container is named "AutoML"
and should be selected for regular use of the AutoML extension.

The purpose of this Dev Container is that it will be used instead of
relying on the Dev Container of a third-party repository.

Signed-off-by: Pedro Zawadniak <Pedro.FontouraZawadniak@analog.com>